test: keep review assertions independent
